requirements.txt
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.10.11
aiosignal==1.3.1
airdrop-ng==1.1
airgraph-ng==1.1
alabaster==1.0.0
alembic==1.13.1
annotated-types==0.7.0
ansible==11.4.0
ansible-core==2.18.4
anyio==4.8.0
apparmor==4.0.3
appdirs==1.4.4
application-utility==1.4.0
arandr==0.1.11
argcomplete==3.5.3
asciidoc==10.2.1
astroid==3.3.9
attrs==23.2.1.dev0
autocommand==2.2.2
Babel==2.15.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
btrfsutil==6.13
build==1.2.2
CacheControl==0.14.2
cachetools==5.5.2
cattrs==24.1.3
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cheroot==10.0.1
CherryPy==18.10.0
ci-info==0.3.0
click==8.1.8
click-plugins==1.1.1
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
coverage==7.6.12
cryptography==44.0.2
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.12
dbus-python==1.4.0
decorator==5.1.1
dill==0.3.9
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docstring-to-markdown==0.15
docutils==0.21.2
docx2pdf==0.1.8
docx2txt==0.9
elastic-transport==8.17.1
elasticsearch==8.17.2
email_validator==2.2.0
entrypoints==0.4
etelemetry==0.3.1
exceptiongroup==1.2.2
extras==1.0.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.18.0
fixtures==4.2.3
flake8==7.1.2
Flask==3.1.0
flask-babel==4.0.0
Flask-Compress==1.17
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Mailman==1.1.1
Flask-Migrate==4.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security==5.6.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flit_core==3.11.0
fonttools==4.57.0
frozenlist==1.4.1
furo==2024.8.6
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-tools==1.71.0
gssapi==1.9.0
gufw==24.4.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
imagesize==1.4.1
importlib_metadata==7.2.1
importlib_resources==6.5.2
inflect==7.5.0
iniconfig==2.0.0
installer==0.7.0
isodate==0.6.1
isort==6.0.1
itsdangerous==2.1.2
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jedi==0.19.2
jedi-language-server==0.44.0
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
keyutils==0.6
kiwisolver==1.4.5
langdetect==1.0.9
ldap3==2.9
legacy-cgi==2.6.3
lensfun==0.3.4
LibAppArmor==4.0.3
libvirt-python==11.2.0
lightdm-gtk-greeter-settings==1.2.3
lit==19.1.7.dev0
lockfile==0.12.2
logutils==0.3.5
looseversion==1.3.0
lsprotocol==2023.0.1
lxml==5.3.1
Mako==1.3.9.dev0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mccabe==0.7.0
mdurl==0.1.2
menulibre==2.4.0
mercurial==7.0
meson==1.7.2
more-itertools==10.6.0
msgpack==1.0.5
mugshot==0.4.3
multidict==6.0.5
neovim-remote==2.5.1
networkx==3.4.2
nftables==0.1
nibabel==5.3.2
nipype==1.10.0
nose==1.3.7
npyscreen==4.10.5
nspektr==0.5.0
numpy==2.2.4
oauthlib==3.2.2
ordered-set==4.1.0
orjson==3.10.16
outcome==1.3.0.post0
packaging==24.2
pacman_mirrors==4.27
pandas==2.2.3
paramiko==3.5.0
parso==0.8.4
passlib==1.7.4
pathlib==1.0.1
pbr==6.1.1
pdf2image==1.17.0
pecan==1.5.1
pexpect==4.9.0
pillow==11.1.0
pipenv==2024.4.1
pipx==1.7.1
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
pmbootstrap==3.3.2
pooch==1.8.2
portend==3.2.0
powerline-status==2.8.4
prettytable==3.15.1
proto-plus==1.26.1
protobuf==6.30.1
prov==2.0.1
psutil==7.0.0
psycopg2==2.9.9
ptyprocess==0.7.0
puremagic==1.29
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.27.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.11.2
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydot==3.0.4
pyflakes==3.2.0
pygdbmi==0.11.0.0
pygls==1.3.1
Pygments==2.19.1
PyGObject==3.52.3
PyJWT==2.10.1
pylint==3.3.5
PyNaCl==1.5.0
pynvim==0.5.2
pyOpenSSL==25.0.0
pyparsing==3.2.1
pypdf==5.4.0
pyproject_hooks==1.2.0
pyserial==3.5
PySide6==6.9.0
pytesseract==0.3.13
pytest==8.3.5
pytest-cov==6.1.0
python-dateutil==2.9.0
python-docx==1.1.2
python-dotenv==1.1.0
python-editor==1.0.4
python-magic==0.4.27
python-multipart==0.0.20
python3-saml==1.16.0
pytz==2025.1
pyxdg==0.28
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
resolvelib==1.1.0
retrying==1.3.4
rich==14.0.0
rich-toolkit==0.14.3
roman-numerals-py==3.1.0
rpm==4.20.1
rsa==4.9
ruff==0.11.4
ruff-lsp==0.0.62
s3cmd==2.4.0
scipy==1.15.2
scour==0.38.2
semantic_version==2.10.0
sentry-sdk==2.25.1
setuptools==75.8.0
setuptools-rust==1.11.1
shellingham==1.5.4
shiboken6==6.9.0
shiboken6-generator==6.9.0
shodan==1.31.0
simplejson==3.20.1
six==1.17.0
smbus==1.1
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
speaklater==1.3
Sphinx==8.2.3
sphinx-copybutton==0.5.2
sphinx_basic_ng==1.0.0b2
sphinx_inline_tabs==2023.4.21
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sphinxext-opengraph==0.9.1
sqlparse==0.5.1
sshtunnel==0.4.0
starlette==0.46.2
systemd-python==235
TBB==0.2
tempora==5.8.0
tenacity==9.1.0
testpath==0.6.0
testtools==2.7.2
tldextract==5.1.3
tlp-ui==1.8.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
tqdm==4.67.1
traits==7.0.2
trio==0.26.2
trio-websocket==0.12.2
trove-classifiers==2025.3.19.19
typeguard==4.4.2
typer==0.15.2
typesense==1.0.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
uc-micro-py==1.0.3
udiskie==2.5.7
ufw==0.36.2
uritemplate==4.1.1
urllib3==2.3.0
userpath==1.9.2
uvicorn==0.34.2
uvloop==0.21.0
validate-pyproject==0.24.1
virtualenv==20.28.0
virtualenv-clone==0.5.7
waitress==3.0.2
watchfiles==1.0.5
wcwidth==0.2.13
webencodings==0.5.1
WebOb==1.8.8
websocket-client==1.8.0
websockets==15.0.1
WebTest==3.0.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
WTForms==3.2.1
XlsxWriter==3.2.2
xmlsec==1.3.15.dev0+g1b3b527.d20241117
yarl==1.13.1
zc.lockfile==3.0.post1
zipp==3.21.0
zstandard==0.23.0
